chatbot/requirements.txt
pinecone-client
openai
langchain
pandas
numpy
python-dotenv
boto3
langchain_pinecone
langchain-community